Speed up pos -> idx lookups
