+++ onoro_impl/src/lib.rs
@@ -12,6 +12,8 @@ mod p1_move_gen;
 mod p2_move_gen;
 mod packed_hex_pos;
 mod packed_idx;
+#[cfg(test)]
+mod test_util;
 mod tile_hash;
 mod util;
