+++ onoro_impl/src/p2_move_gen.rs
@@ -113,12 +113,7 @@ impl<const N: usize> P2MoveGenerator<N> {
       .neighbors(pawn_index as usize)
       .map(|neighbor_index| {
         let neighbor_pos = indexer.pos_from_index(neighbor_index);
-        pawn_poses
-          .iter()
-          .enumerate()
-          .find(|&(_, &pos)| pos == neighbor_pos)
-          .unwrap()
-          .0
+        OnoroImpl::pawn_idx_from_pawn_poses(pawn_poses, neighbor_pos) as usize
       })
   }
